+++ omni_lang/omni_lang/tests/test63.nim
@@ -1,0 +1,11 @@
+import ../../omni_lang
+
+struct Something:
+    data Data
+
+init:
+    something = Something(Data())
+    bubu := something.data
+
+sample:
+    out1 = bubu[0]